fixed sorting by `normalized`
Sorting by normalized did not work propperly since it is sorted
lexically. Shorter values (aka with less decimal points) always floated
to the top. Since I can't (or don't want) to fix the sorting of the
data-tables I mittigate the issue by formating these values to 4 fixed
decimal points. We can assume the values to be always smaller than 1.27
(half of that normalized inch in cm) so we don't really have to worry
about pre-decimal points.